Moved the pretty writer code from slang-reflection-test into core
