+++ tools/slang-reflection-test/slang-reflection-test-main.cpp
@@ -4,6 +4,7 @@
 #include "../../source/core/slang-string-escape-util.h"
 #include "../../source/core/slang-string-util.h"
 #include "../../source/core/slang-test-tool-util.h"
+#include "../../source/core/slang-pretty-writer.h"
 #include "slang-com-helper.h"
 #include "slang.h"
 
@@ -60,190 +61,6 @@ Range<T> makeRange(T end)
     return Range<T>(T(0), end);
 }
 
-
-struct PrettyWriter
-{
-    typedef PrettyWriter ThisType;
-
-    friend struct CommaTrackerRAII;
-
-    struct CommaState
-    {
-        bool needComma = false;
-    };
-
-    void writeRaw(const UnownedStringSlice& slice) { m_builder.append(slice); }
-    void writeRaw(char const* begin, char const* end);
-    void writeRaw(PrettyWriter& writer, char const* begin) { writeRaw(UnownedStringSlice(begin)); }
-
-    void writeRawChar(int c) { m_builder.appendChar(char(c)); }
-
-    void writeHexChar(int c) { writeRawChar(CharUtil::getHexChar(Index(c))); }
-
-    /// Adjusts indentation if at start of a line
-    void adjust();
-
-    /// Increase indentation
-    void indent() { m_indent++; }
-    /// Decreate indentation
-    void dedent();
-
-    /// Write taking into account any CR that might be in a slice
-    void write(const UnownedStringSlice& slice);
-    void write(char const* text) { write(UnownedStringSlice(text)); }
-    void write(char const* text, size_t length) { write(UnownedStringSlice(text, length)); }
-
-    /// Write the slice as an escaped string
-    void writeEscapedString(const UnownedStringSlice& slice);
-
-    /// Call before items in a comma-separated JSON list to emit the comma if/when needed
-    void maybeComma();
-
-    /// Get the builder the result is being constructed in
-    StringBuilder& getBuilder() { return m_builder; }
-
-    ThisType& operator<<(const UnownedStringSlice& slice)
-    {
-        write(slice);
-        return *this;
-    }
-    ThisType& operator<<(const char* text)
-    {
-        write(text);
-        return *this;
-    }
-    ThisType& operator<<(uint64_t val)
-    {
-        adjust();
-        m_builder << val;
-        return *this;
-    }
-    ThisType& operator<<(int64_t val)
-    {
-        adjust();
-        m_builder << val;
-        return *this;
-    }
-    ThisType& operator<<(int32_t val)
-    {
-        adjust();
-        m_builder << val;
-        return *this;
-    }
-    ThisType& operator<<(uint32_t val)
-    {
-        adjust();
-        m_builder << val;
-        return *this;
-    }
-    ThisType& operator<<(float val)
-    {
-        adjust();
-        // We want to use a specific format, so we use the StringUtil to specify format, and not
-        // just use <<
-        StringUtil::appendFormat(m_builder, "%f", val);
-        return *this;
-    }
-
-    bool m_startOfLine = true;
-    int m_indent = 0;
-    CommaState* m_commaState = nullptr;
-    StringBuilder m_builder;
-};
-
-void PrettyWriter::writeRaw(char const* begin, char const* end)
-{
-    SLANG_ASSERT(end >= begin);
-    writeRaw(UnownedStringSlice(begin, end));
-}
-
-void PrettyWriter::adjust()
-{
-    // Only indent if at start of a line
-    if (m_startOfLine)
-    {
-        // Output current indentation
-        m_builder.appendRepeatedChar(' ', m_indent * 4);
-        m_startOfLine = false;
-    }
-}
-
-void PrettyWriter::dedent()
-{
-    SLANG_ASSERT(m_indent > 0);
-    m_indent--;
-}
-
-void PrettyWriter::write(const UnownedStringSlice& slice)
-{
-    const auto end = slice.end();
-    auto start = slice.begin();
-
-    while (start < end)
-    {
-        const char* cur = start;
-
-        // Search for \n if there is one
-        while (cur < end && *cur != '\n')
-            cur++;
-
-        // If there were some chars, adjust and write
-        if (cur > start)
-        {
-            adjust();
-            writeRaw(UnownedStringSlice(start, cur));
-        }
-
-        if (cur < end && *cur == '\n')
-        {
-            writeRawChar('\n');
-            // Skip the CR
-            cur++;
-            // Mark we are at the start of a line
-            m_startOfLine = true;
-        }
-
-        start = cur;
-    }
-}
-
-void PrettyWriter::writeEscapedString(const UnownedStringSlice& slice)
-{
-    adjust();
-    auto handler = StringEscapeUtil::getHandler(StringEscapeUtil::Style::Cpp);
-    StringEscapeUtil::appendQuoted(handler, slice, m_builder);
-}
-
-void PrettyWriter::maybeComma()
-{
-    if (auto state = m_commaState)
-    {
-        if (!state->needComma)
-        {
-            state->needComma = true;
-            return;
-        }
-    }
-
-    write(toSlice(",\n"));
-}
-
-/// Type for tracking whether a comma is needed in a comma-separated JSON list
-struct CommaTrackerRAII
-{
-    CommaTrackerRAII(PrettyWriter& writer)
-        : m_writer(&writer), m_previousState(writer.m_commaState)
-    {
-        writer.m_commaState = &m_state;
-    }
-
-    ~CommaTrackerRAII() { m_writer->m_commaState = m_previousState; }
-
-private:
-    PrettyWriter::CommaState m_state;
-    PrettyWriter* m_writer;
-    PrettyWriter::CommaState* m_previousState;
-};
 
 
 static void emitReflectionVarInfoJSON(PrettyWriter& writer, slang::VariableReflection* var);
